add fiture message broker, chacing, upload

+++ src/services/postgres/PlaylistSongsService.js
@@ -1,18 +1,18 @@
 const { Pool } = require('pg');
 const { nanoid } = require('nanoid');
-const { mapDBToModel } = require('../../utils');
 
 const InvariantError = require('../../exceptions/InvariantError');
 const NotFoundError = require('../../exceptions/NotFoundError');
 const AuthorizationError = require('../../exceptions/AuthorizationError');
 
 class PlaylistSongService {
-  constructor(collaborationsService) {
+  constructor(collaborationsService, cacheService) {
     this._pool = new Pool();
     this._collaborationsService = collaborationsService;
+    this._cacheService = cacheService;
   }
 
-  async addPlaylistSong(playlist_Id, songId) {
+  async addPlaylistSong(playlist_Id, songId, owner) {
     const id = `PlaylistSong-${nanoid(16)}`;
     const query = {
       text: 'INSERT INTO playlistsongs VALUES($1, $2, $3) RETURNING id',
@@ -23,25 +23,32 @@ class PlaylistSongService {
     if (!result.rowCount) {
       throw new InvariantError('Lagu gagal ditambahkan ke playlist');
     }
+    await this._cacheService.delete(`songs:${owner}`);
     return result.rows[0].id;
   }
 
   async getPlaylistSongs(owner) {
-    const query = {
-      text: 'select songs.id, songs.title, songs.performer from playlistsongs full join songs on playlistsongs.song_id = songs.id full join playlists on playlistsongs.playlist_id = playlists.id full join collaborations on collaborations.playlist_id=playlists.id where collaborations.user_id =  $1 or playlists.owner = $1',
-      values: [owner],
-    };
-
-    const result = await this._pool.query(query);
-    if (!result.rowCount) {
-      throw new NotFoundError('lagu tidak ditemukan di dalam playlists');
+    try {
+      const result = await this._cacheService.get(`songs:${owner}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: 'select songs.id, songs.title, songs.performer from playlistsongs full join songs on playlistsongs.song_id = songs.id full join playlists on playlistsongs.playlist_id = playlists.id full join collaborations on collaborations.playlist_id=playlists.id where collaborations.user_id =  $1 or playlists.owner = $1',
+        values: [owner],
+      };
+      const result = await this._pool.query(query);
+      if (!result.rowCount) {
+        throw new NotFoundError('lagu tidak ditemukan di dalam playlists');
+      }
+      const thefinalresult = result.rows;
+      await this._cacheService.set(`songs:${owner}`, JSON.stringify(thefinalresult));
+      return thefinalresult;
     }
-    return result.rows.map(mapDBToModel);
   }
 
   async deletePlaylistSongsById(playlist_Id) {
     const query = {
-      text: 'DELETE FROM playlistsongs WHERE playlist_id = $1 RETURNING id',
+      text: 'DELETE FROM playlistsongs USING playlists WHERE playlists.id = $1 and playlistsongs.playlist_id = $1 RETURNING playlistsongs.id, playlists.owner',
       values: [playlist_Id],
     };
 
@@ -50,6 +57,8 @@ class PlaylistSongService {
     if (!result.rowCount) {
       throw new InvariantError('id tidak valid');
     }
+    const { owner } = result.rows[0];
+    await this._cacheService.delete(`songs:${owner}`);
   }
 
   async verifyPlaylistSongsOwner(playlist_id, user_id) {
